chore: The LogEntryPreviewModal component was refactored, date-picker related codes were extracted out into a new LogEntryDatePicker component. The purpose of this change is to improve readability and maintenance of the codebase, while retaining original features for users.
