+++ packages/apps/spaces/app/organizations/[id]/components/Timeline/events/logEntry/LogEntryUpdateFormDto.ts
@@ -27,11 +27,10 @@ export class LogEntryUpdateFormDto implements LogEntryUpdateForm {
     date: Date | string,
     time: string,
   ): Date {
-    const timeArray = time.split(':');
+    const timeArray = time?.split(':');
     const newDate = new Date(date); // Create a new Date object to maintain immutability
     newDate.setHours(Number(timeArray?.[0] ?? '00'));
-    newDate.setMinutes(Number(timeArray[1]?.[0] ?? '00'));
-
+    newDate.setMinutes(Number(timeArray?.[1] ?? '00'));
     return newDate;
   }
   static toPayload(
